+++ philo/f_threads_create.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   f_threads_create.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: a600 <a600@student.42.fr>                  +#+  +:+       +#+        */
+/*   By: fsitter <fsitter@student.42vienna.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/27 16:18:23 by fsitter           #+#    #+#             */
-/*   Updated: 2026/01/29 23:51:15 by a600             ###   ########.fr       */
+/*   Updated: 2026/01/30 13:02:16 by fsitter          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,7 +46,7 @@ int	f_init_threads(t_philo **ph, pthread_t *threads)
 
 void	f_join_threads(t_in *in, pthread_t *threads)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < in->nop + 1)
